Fixed phpstan errors
